requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pymongo==4.10.1
jinja2==3.1.4
python-multipart==0.0.20
pydantic==2.10.3
python-dotenv==1.0.1
